feat: Add endpoint for ending subscription
This commit adds a new endpoint `/api/clusters/{clusterName}/sales/end-subscription` to the server API. It also includes corresponding changes to the `clusterSlice.js` file in the frontend code.

The new endpoint allows users to end their subscription to a cluster. The `endSubscription` function in the frontend code sends a request to the server with the cluster name and username. If the request is successful, a success banner is shown, and the subscription is ended. If there is an error, an error banner is shown.

This feature was implemented to provide users with the ability to easily end their subscription to a cluster.